Meta annotations are now imported
Meta annotations are now implicitely imported if the corresponding meta annotated class is
explicitely imported. Before that, this was only true for direct
annotations. This is done the same way as for class
hierarchies: After the import of all direct annotations, it is checked
whether each of those annotations has a annotation itself which is then
imported. Since this process is done recursively, the whole hierarchy of annotations is
imported, i.e., annotations of meta annotations and so on. In addition
to that, parameter of annotations are imported as well.

Signed-off-by: Korbinian Singhammer <korbinian.singhammer@tngtech.com>