Meta annotations of members are imported
If a class with a meta annotated member is explicitely imported, the
corresponding meta annotation, togehter with its parameters, is imported
as well. Before, only direct annotations of members were imported. Now,
once such a direct annotatation is imported, an iterator goes through
its annotation (which is the meta annotation of the member) and imports
it.

Signed-off-by: Korbinian Singhammer <korbinian.singhammer@tngtech.com>